refactor: preprocessing into own directory while encapsulating logic into some classes
