+++ include/silo/preprocessing/preprocessing_exception.h
@@ -3,11 +3,11 @@
 #include <stdexcept>
 #include <string>
 
-namespace silo {
+namespace silo::preprocessing {
 
 class PreprocessingException : public std::runtime_error {
   public:
    explicit PreprocessingException(const std::string& error_message);
 };
 
-}  // namespace silo
+}  // namespace silo::preprocessing
