+++ src/silo/database.cpp
@@ -43,11 +43,8 @@
 #include "silo/config/database_config.h"
 #include "silo/database_info.h"
 #include "silo/persistence/exception.h"
-#include "silo/prepare_dataset.h"
 #include "silo/preprocessing/metadata_validator.h"
 #include "silo/preprocessing/ndjson_digestion.h"
-#include "silo/preprocessing/pango_lineage_count.h"
-#include "silo/preprocessing/partition.h"
 #include "silo/preprocessing/preprocessing_config.h"
 #include "silo/preprocessing/preprocessing_exception.h"
 #include "silo/query_engine/query_engine.h"
@@ -69,20 +66,6 @@
 #include "silo/zstdfasta/zstd_decompressor.h"
 #include "silo/zstdfasta/zstdfasta_table.h"
 #include "silo/zstdfasta/zstdfasta_table_reader.h"
-
-template <>
-struct [[maybe_unused]] fmt::formatter<silo::DatabaseInfo> : fmt::formatter<std::string> {
-   [[maybe_unused]] static auto format(silo::DatabaseInfo database_info, format_context& ctx)
-      -> decltype(ctx.out()) {
-      return format_to(
-         ctx.out(),
-         "sequence count: {}, total size: {}, N bitmaps size: {}",
-         database_info.sequence_count,
-         silo::formatNumber(database_info.total_size),
-         silo::formatNumber(database_info.n_bitmaps_size)
-      );
-   }
-};
 
 namespace silo {
 
@@ -124,99 +107,6 @@ template <>
 const std::map<std::string, SequenceStore<AminoAcid>>& Database::getSequenceStores<AminoAcid>(
 ) const {
    return aa_sequences;
-}
-
-const PangoLineageAliasLookup& Database::getAliasKey() const {
-   return alias_key;
-}
-
-void Database::build(
-   duckdb::Connection& connection,
-   const preprocessing::Partitions& partition_descriptor,
-   const ReferenceGenomes& reference_genomes,
-   const std::string& order_by_clause
-) {
-   int64_t micros = 0;
-   {
-      const BlockTimer timer(micros);
-      for (const auto& partition : partition_descriptor.getPartitions()) {
-         partitions.emplace_back(partition.getPartitionChunks());
-      }
-      initializeColumns();
-      initializeNucSequences(reference_genomes.nucleotide_sequences);
-      initializeAASequences(reference_genomes.aa_sequences);
-
-      SPDLOG_INFO("build - building metadata store");
-
-      for (size_t partition_id = 0; partition_id < partition_descriptor.getPartitions().size();
-           ++partition_id) {
-         const auto& part = partition_descriptor.getPartitions()[partition_id];
-         for (size_t chunk_index = 0; chunk_index < part.getPartitionChunks().size();
-              ++chunk_index) {
-            partitions[partition_id].sequence_count += partitions[partition_id].columns.fill(
-               connection, partition_id, order_by_clause, database_config
-            );
-         }
-         SPDLOG_INFO("build - finished columns for partition {}", partition_id);
-      }
-
-      SPDLOG_INFO("build - building sequence stores");
-
-      tbb::parallel_for(
-         tbb::blocked_range<size_t>(0, partition_descriptor.getPartitions().size()),
-         [&](const auto& local) {
-            for (auto partition_index = local.begin(); partition_index != local.end();
-                 ++partition_index) {
-               const auto& part = partition_descriptor.getPartitions()[partition_index];
-               for (size_t chunk_index = 0; chunk_index < part.getPartitionChunks().size();
-                    ++chunk_index) {
-                  for (const auto& [nuc_name, reference_sequence] :
-                       reference_genomes.raw_nucleotide_sequences) {
-                     SPDLOG_DEBUG(
-                        "build - building sequence store for nucleotide sequence {} and partition "
-                        "{}",
-                        nuc_name,
-                        partition_index
-                     );
-
-                     silo::ZstdFastaTableReader sequence_input(
-                        connection,
-                        "nuc_" + nuc_name,
-                        reference_sequence,
-                        fmt::format("partition_id = {}", partition_index),
-                        order_by_clause
-                     );
-                     partitions[partition_index].nuc_sequences.at(nuc_name).fill(sequence_input);
-                  }
-                  for (const auto& [aa_name, reference_sequence] :
-                       reference_genomes.raw_aa_sequences) {
-                     SPDLOG_DEBUG(
-                        "build - building sequence store for amino acid sequence {} and partition "
-                        "{}",
-                        aa_name,
-                        partition_index
-                     );
-
-                     silo::ZstdFastaTableReader sequence_input(
-                        connection,
-                        "gene_" + aa_name,
-                        reference_sequence,
-                        fmt::format("partition_id = {}", partition_index),
-                        order_by_clause
-                     );
-                     partitions[partition_index].aa_sequences.at(aa_name).fill(sequence_input);
-                  }
-               }
-               partitions.at(partition_index).flipBitmaps();
-               SPDLOG_INFO("build - finished sequences for partition {}", partition_index);
-            }
-         }
-      );
-      finalizeInsertionIndexes();
-   }
-
-   SPDLOG_INFO("Build took {} ms", micros);
-   SPDLOG_INFO("database info: {}", getDatabaseInfo());
 }
 
 using RoaringStatistics = roaring::api::roaring_statistics_t;
@@ -340,7 +230,6 @@ void addStatisticToBitmapContainerSize(
    const RoaringStatistics& statistic,
    BitmapContainerSizeStatistic& size_statistic
 ) {
-   SPDLOG_TRACE("addStatisticToBitmapContainerSize");
    size_statistic.number_of_array_containers += statistic.n_array_containers;
    size_statistic.number_of_run_containers += statistic.n_run_containers;
    size_statistic.number_of_bitset_containers += statistic.n_bitset_containers;
@@ -647,400 +536,6 @@ Database Database::loadDatabaseState(const std::filesystem::path& save_directory
    SPDLOG_INFO(
       "Finished loading data_version from {}", (save_directory / "data_version.silo").string()
    );
-
-   return database;
-}
-
-namespace {
-
-preprocessing::Partitions buildPartitionDescriptorFromSQL(duckdb::Connection& connection) {
-   auto partition_descriptor_from_sql =
-      connection.Query("SELECT partition_id, count FROM partitioning ORDER BY partition_id");
-
-   std::vector<preprocessing::Partition> partitions;
-
-   uint32_t check_partition_id_sorted_and_contiguous = 0;
-
-   for (auto it = partition_descriptor_from_sql->begin();
-        it != partition_descriptor_from_sql->end();
-        ++it) {
-      const duckdb::Value db_partition_id = it.current_row.GetValue<duckdb::Value>(0);
-      int64_t partition_id_int = duckdb::BigIntValue::Get(db_partition_id);
-      if (partition_id_int != check_partition_id_sorted_and_contiguous) {
-         throw silo::PreprocessingException(
-            "The partition IDs produced by the preprocessing are not sorted, not starting from 0 "
-            "or not contiguous."
-         );
-      }
-      check_partition_id_sorted_and_contiguous++;
-      uint32_t partition_id = partition_id_int;
-
-      const duckdb::Value db_partition_size = it.current_row.GetValue<duckdb::Value>(1);
-      int64_t partition_size_bigint = duckdb::BigIntValue::Get(db_partition_size);
-      if (partition_size_bigint <= 0) {
-         throw silo::PreprocessingException("Non-positive partition size encountered.");
-      }
-      if (partition_size_bigint > UINT32_MAX) {
-         throw silo::PreprocessingException(
-            fmt::format("Overflow of limit UINT32_MAX ({}) for number of sequences.", UINT32_MAX)
-         );
-      }
-
-      uint32_t partition_size = static_cast<uint32_t>(partition_size_bigint);
-
-      partitions.emplace_back(std::vector<preprocessing::PartitionChunk>{
-         {partition_id, 0, partition_size, 0}});
-   }
-
-   return preprocessing::Partitions(partitions);
-}
-
-}  // namespace
-
-Database Database::preprocessing(
-   const preprocessing::PreprocessingConfig& preprocessing_config,
-   const config::DatabaseConfig& database_config_
-) {
-   Database database;
-   database.database_config = database_config_;
-
-   const DataVersion& data_version = DataVersion::mineDataVersion();
-   SPDLOG_INFO("preprocessing - mining data data_version: {}", data_version.toString());
-   database.setDataVersion(data_version);
-
-   SPDLOG_INFO("preprocessing - reading reference genome");
-   const ReferenceGenomes& reference_genomes =
-      ReferenceGenomes::readFromFile(preprocessing_config.getReferenceGenomeFilename());
-
-   const std::optional<std::string> ndjson_input_filename =
-      preprocessing_config.getNdjsonInputFilename();
-
-   duckdb::DuckDB preprocessing_memory(
-      preprocessing_config.getOutputDirectory() / "preprocessing.duckdb"
-   );  // TODO make configurable via preprocessing config
-   duckdb::Connection connection(preprocessing_memory);
-   {
-      auto return_code = connection.Query("PRAGMA default_null_order='NULLS FIRST';");
-      if (return_code->HasError()) {
-         throw silo::PreprocessingException(return_code->GetError());
-      }
-   }
-
-   std::string order_by_clause;  // TODO turn into struct
-   if (database_config_.schema.date_to_sort_by.has_value()) {
-      SPDLOG_INFO("preprocessing - produce order by clause with a date to sort by");
-
-      order_by_clause = fmt::format(
-         "ORDER BY {}, {}",
-         database_config_.schema.date_to_sort_by.value(),
-         database_config_.schema.primary_key
-      );
-   } else {
-      SPDLOG_INFO("preprocessing - produce order by clause without a date to sort by");
-
-      order_by_clause = fmt::format("ORDER BY {}", database_config_.schema.primary_key);
-   }
-   SPDLOG_INFO("preprocessing - order by clause is {}", order_by_clause);
-
-   if (ndjson_input_filename.has_value()) {
-      SPDLOG_DEBUG("preprocessing - ndjson pipeline chosen");
-
-      executeDuckDBRoutineForNdjsonDigestion(
-         connection,
-         preprocessing_config,
-         reference_genomes,
-         ndjson_input_filename.value(),
-         database.database_config.schema.primary_key
-      );
-
-      auto return_code = connection.Query(
-         "create or replace view metadata_table as\n"
-         "select metadata.*, insertions, aaInsertions\n"
-         "from preprocessing_table;"
-      );
-      if (return_code->HasError()) {
-         SPDLOG_ERROR(return_code->GetError());
-         throw silo::PreprocessingException(return_code->GetError());
-      }
-
-      SPDLOG_INFO("preprocessing - finished building the in-memory table for preprocessing");
-      const std::string peek_query = "SELECT * FROM preprocessing_table LIMIT 5;";
-      SPDLOG_TRACE(
-         "preprocessing - peek into the table: {} \n {}",
-         peek_query,
-         connection.Query(peek_query)->ToString()
-      );
-   } else {
-      SPDLOG_DEBUG("preprocessing - classic pipeline chosen");
-
-      auto return_code = connection.Query(fmt::format(
-         "create or replace table metadata_table as\n"
-         "select *\n"
-         "from '{}';",
-         preprocessing_config.getMetadataInputFilename().string()
-      ));
-      if (return_code->HasError()) {
-         SPDLOG_ERROR(return_code->GetError());
-         throw silo::PreprocessingException(return_code->GetError());
-      }
-   }
-
-   // TODO SPDLOG_INFO("preprocessing - validate metadata file (-> preprocessing_table) against
-   // config");
-   // TODO also validate types
-
-   const std::string metadata_filename = preprocessing_config.getMetadataInputFilename().string();
-   preprocessing::MetadataValidator().validateMedataFile(
-      preprocessing_config.getMetadataInputFilename(), database_config_
-   );
-
-   SPDLOG_INFO("preprocessing - building alias key");
-   const auto pango_lineage_definition_filename =
-      preprocessing_config.getPangoLineageDefinitionFilename();
-   if (pango_lineage_definition_filename.has_value()) {
-      database.alias_key =
-         PangoLineageAliasLookup::readFromFile(pango_lineage_definition_filename.value());
-   }
-
-   if (database_config_.schema.partition_by.has_value()) {
-      SPDLOG_INFO("preprocessing - calculating partitions");
-
-      auto return_code = connection.Query(fmt::format(
-         "create\n"
-         "or replace table partition_keys as\n"
-         "select row_number() over () - 1 as id, *\n"
-         "from (SELECT {} as partition_key, COUNT(*) as count "
-         "      FROM metadata_table "
-         "      GROUP BY partition_key "
-         "      ORDER BY partition_key);",
-         database_config_.schema.partition_by.value()  // TODO make sure partition key is validated
-      ));
-
-      if (return_code->HasError()) {
-         throw silo::PreprocessingException(
-            "Error in the execution of the duckdb statement for partition key table "
-            "generation: " +
-            return_code->GetError()
-         );
-      } else {
-         SPDLOG_TRACE("Executed statement for partition key table generation.");
-         SPDLOG_TRACE(return_code->ToString());
-      }
-
-      return_code = connection.Query(
-         "create or replace table partitioning as\n"
-         "with recursive "
-         "          allowed_count(allowed_count) as (select sum(count) / 32 from "
-         "partition_keys),\n"
-         "          grouped_partition_keys(from_id, to_id, count) as\n"
-         "              (select id, id, count\n"
-         "               from partition_keys\n"
-         "               where id = 0\n"
-         "               union all\n"
-         "               select case when l1.count <= allowed_count then l1.from_id else l2.id end,"
-         "                      l2.id,\n"
-         "                      case when l1.count <= allowed_count\n"
-         "                           then l1.count + l2.count\n"
-         "                           else l2.count end\n"
-         "               from grouped_partition_keys l1,\n"
-         "                    partition_keys l2,\n"
-         "                    allowed_count\n"
-         "               where l1.to_id + 1 = l2.id)\n"
-         "select row_number() over () - 1 as partition_id, *\n"
-         "from (select from_id, max(to_id) as to_id, max(count) as count\n"
-         "      from grouped_partition_keys\n"
-         "      group by from_id)"
-      );
-      if (return_code->HasError()) {
-         SPDLOG_ERROR("Error when executing duckdb statement: {}", return_code->GetError());
-         throw silo::PreprocessingException(return_code->GetError());
-      }
-
-      return_code = connection.Query(
-         "create\n"
-         "or replace table partition_key_to_partition as\n"
-         "select partition_keys.partition_key as partition_key, "
-         "  partitioning.partition_id as partition_id\n"
-         "from partition_keys,\n"
-         "     partitioning\n"
-         "where partition_keys.id >= partitioning.from_id\n"
-         "  AND partition_keys.id <= partitioning.to_id;"
-      );
-      if (return_code->HasError()) {
-         SPDLOG_ERROR("Error when executing duckdb statement: {}", return_code->GetError());
-         throw PreprocessingException(return_code->GetError());
-      }
-
-      return_code = connection.Query(fmt::format(
-         "create\n"
-         "or replace view partitioned_metadata as\n"
-         "select partitioning.partition_id as partition_id, metadata_table.*\n"
-         "from partition_keys,\n"
-         "     partitioning,\n"
-         "     metadata_table\n"
-         "where (metadata_table.{0} = partition_keys.partition_key or (metadata_table.{0} is null "
-         "and partition_keys.partition_key is null))\n"
-         "  AND partition_keys.id >= partitioning.from_id\n"
-         "  AND partition_keys.id <= partitioning.to_id;",
-         database_config_.schema.partition_by.value()
-      ));
-      if (return_code->HasError()) {
-         SPDLOG_ERROR("Error when executing duckdb statement: {}", return_code->GetError());
-         throw PreprocessingException(return_code->GetError());
-      }
-   } else {
-      SPDLOG_INFO(
-         "preprocessing - skip partition merging because no partition_by key was provided, instead "
-         "putting all sequences into the same partition"
-      );
-
-      auto return_code = connection.Query(
-         "create or replace table partitioning as\n"
-         "select 0::bigint as partition_id, 0::bigint as from_id, 0::bigint as to_id, count(*) as "
-         "count\n"
-         "from metadata_table;"
-      );
-
-      if (return_code->HasError()) {
-         SPDLOG_ERROR(return_code->GetError());
-         throw silo::PreprocessingException(return_code->GetError());
-      }
-
-      return_code = connection.Query(
-         "create\n"
-         "or replace table partition_key_to_partition as\n"
-         "select 0::bigint as partition_key, 0::bigint as partition_id;"
-      );
-
-      if (return_code->HasError()) {
-         SPDLOG_ERROR(return_code->GetError());
-         throw silo::PreprocessingException(return_code->GetError());
-      }
-
-      return_code = connection.Query(
-         "create\n"
-         "or replace view partitioned_metadata as\n"
-         "select 0::bigint as partition_id, metadata_table.*\n"
-         "from metadata_table;"
-      );
-
-      if (return_code->HasError()) {
-         SPDLOG_ERROR(return_code->GetError());
-         throw silo::PreprocessingException(return_code->GetError());
-      }
-   }
-
-   if (preprocessing_config.getNdjsonInputFilename().has_value()) {
-      std::string order_by_select = ", metadata." + database_config_.schema.primary_key + " as " +
-                                    database_config_.schema.primary_key;
-      if (database_config_.schema.date_to_sort_by.has_value()) {
-         order_by_select += ", metadata." + database_config_.schema.date_to_sort_by.value() +
-                            " as " + database_config_.schema.date_to_sort_by.value();
-      }
-
-      for (const auto& [seq_name, _] : reference_genomes.raw_nucleotide_sequences) {
-         auto return_code = connection.Query(fmt::format(
-            "create or replace view nuc_{0} as\n"
-            "select metadata.{1} as key, nuc_{0} as sequence,"
-            "partition_key_to_partition.partition_id as partition_id"
-            "{3} \n"
-            "from preprocessing_table, partition_key_to_partition "
-            "where preprocessing_table.metadata.{2} = partition_key_to_partition.partition_key;",
-            seq_name,
-            database_config_.schema.primary_key,
-            database_config_.schema.partition_by.value(),
-            order_by_select
-         ));
-
-         if (return_code->HasError()) {
-            SPDLOG_ERROR(return_code->GetError());
-            throw silo::PreprocessingException(return_code->GetError());
-         }
-      }
-
-      for (const auto& [seq_name, _] : reference_genomes.raw_aa_sequences) {
-         auto return_code = connection.Query(fmt::format(
-            "create or replace view gene_{0} as\n"
-            "select metadata.{1} as key, gene_{0} as sequence, "
-            "partition_key_to_partition.partition_id as partition_id\n"
-            "{3} \n"
-            "from preprocessing_table, partition_key_to_partition "
-            "where preprocessing_table.metadata.{2} = partition_key_to_partition.partition_key;",
-            seq_name,
-            database_config_.schema.primary_key,
-            database_config_.schema.partition_by.value(),
-            order_by_select
-         ));
-
-         if (return_code->HasError()) {
-            SPDLOG_ERROR(return_code->GetError());
-            throw silo::PreprocessingException(return_code->GetError());
-         }
-      }
-   } else {
-      std::string order_by_select = ", partitioned_metadata." +
-                                    database_config_.schema.primary_key + " as " +
-                                    database_config_.schema.primary_key;
-      if (database_config_.schema.date_to_sort_by.has_value()) {
-         order_by_select += ", partitioned_metadata." +
-                            database_config_.schema.date_to_sort_by.value() + " as " +
-                            database_config_.schema.date_to_sort_by.value();
-      }
-
-      for (const auto& [seq_name, reference_sequence] :
-           reference_genomes.raw_nucleotide_sequences) {
-         silo::FastaReader fasta_reader(preprocessing_config.getNucFilenameNoExtension(seq_name)
-                                           .replace_extension(silo::preprocessing::FASTA_EXTENSION)
-         );
-         ZstdFastaTable::generate(
-            connection, "raw_nuc_" + seq_name, fasta_reader, reference_sequence
-         );
-
-         auto return_code = connection.Query(fmt::format(
-            "create or replace view nuc_{0} as "
-            "select key, sequence,"
-            "partitioned_metadata.partition_id as partition_id "
-            "{2} \n"
-            "from raw_nuc_{0} raw, partitioned_metadata "
-            "where raw.key = partitioned_metadata.{1};",
-            seq_name,
-            database_config_.schema.primary_key,
-            order_by_select
-         ));
-
-         if (return_code->HasError()) {
-            SPDLOG_ERROR(return_code->GetError());
-            throw silo::PreprocessingException(return_code->GetError());
-         }
-      }
-
-      for (const auto& [seq_name, reference_sequence] : reference_genomes.raw_aa_sequences) {
-         silo::FastaReader fasta_reader(preprocessing_config.getGeneFilenameNoExtension(seq_name)
-                                           .replace_extension(silo::preprocessing::FASTA_EXTENSION)
-         );
-         ZstdFastaTable::generate(
-            connection, "raw_gene_" + seq_name, fasta_reader, reference_sequence
-         );
-
-         auto return_code = connection.Query(fmt::format(
-            "create or replace view gene_{0} as "
-            "select key, sequence,"
-            "partitioned_metadata.partition_id as partition_id"
-            "{2}\n"
-            "from raw_gene_{0} raw, partitioned_metadata "
-            "where raw.key = partitioned_metadata.{1};",
-            seq_name,
-            database_config_.schema.primary_key,
-            order_by_select
-         ));
-      }
-   }
-
-   preprocessing::Partitions partition_descriptor = buildPartitionDescriptorFromSQL(connection);
-
-   SPDLOG_INFO("preprocessing - building database");
-
-   database.build(connection, partition_descriptor, reference_genomes, order_by_clause);
 
    return database;
 }
