+++ src/silo/database.test.cpp
@@ -1,14 +1,18 @@
 #include "silo/database.h"
 
-#include "filesystem"
+#include <filesystem>
+#include <fstream>
 
 #include <gtest/gtest.h>
+#include <nlohmann/json.hpp>
 
 #include "silo/common/nucleotide_symbols.h"
 #include "silo/config/config_repository.h"
 #include "silo/database_info.h"
 #include "silo/preprocessing/preprocessing_config.h"
 #include "silo/preprocessing/preprocessing_config_reader.h"
+#include "silo/preprocessing/preprocessor.h"
+#include "silo/query_engine/query_engine.h"
 
 silo::Database buildTestDatabase() {
    const silo::preprocessing::InputDirectory input_directory{"./testBaseData/exampleDataset/"};
@@ -21,7 +25,18 @@ silo::Database buildTestDatabase() {
       input_directory.directory + "database_config.yaml"
    );
 
-   return silo::Database::preprocessing(config, database_config);
+   silo::preprocessing::Preprocessor preprocessor(config, database_config);
+   return preprocessor.preprocess();
+}
+
+void testSingleQuery(const silo::Database& database, const std::string& query_name) {
+   silo::query_engine::QueryEngine query_engine(database);
+   std::filesystem::path query_directory("endToEndTests/test/queries/");
+   std::ifstream query_file(query_directory / query_name);
+   std::stringstream query_and_result;
+   query_and_result << query_file.rdbuf();
+   nlohmann::json json = nlohmann::json::parse(query_and_result.str());
+   query_engine.executeQuery(json.at("query").dump());
 }
 
 TEST(DatabaseTest, shouldBuildDatabaseWithoutErrors) {
@@ -44,7 +59,8 @@ TEST(DatabaseTest, shouldSuccessfullyBuildDatabaseWithoutPartitionBy) {
       input_directory.directory + "test_database_config_without_partition_by.yaml"
    );
 
-   auto database = silo::Database::preprocessing(config, database_config);
+   silo::preprocessing::Preprocessor preprocessor(config, database_config);
+   auto database = preprocessor.preprocess();
 
    const auto simple_database_info = database.getDatabaseInfo();
